Add animation for SubPages
