+++ src/lib/mobileDetect.ts
@@ -1,0 +1,11 @@
+import { useEffect, useState } from 'react';
+
+export default function useMobileDetect() {
+  const [isMobile, setIsMobile] = useState(false);
+
+  useEffect(() => {
+    setIsMobile(/iPhone|iPad|iPod|Android/i.test(navigator.userAgent));
+  }, []);
+
+  return isMobile;
+}
